third_party/WebKit/LayoutTests/http/tests/security/powerfulFeatureRestrictions/budget-service-on-insecure-origin.html

<!DOCTYPE html>
<title>BudgetService on insecure origin</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/get-host-info.js"></script>
<script>
if (window.location.origin != get_host_info().UNAUTHENTICATED_ORIGIN) {
  window.location = get_host_info().UNAUTHENTICATED_ORIGIN + window.location.pathname;
} else {
  test(t => {
    assert_false(isSecureContext);
    assert_false('getCost' in navigator.budget);
    assert_false('getBudget' in navigator.budget);
    assert_false('reserve' in navigator.budget);
  }, 'navigator.budget requires a secure context');
}
</script>
